[WIP] clean up TIMM skip list
[ghstack-poisoned]
